button on register page add
