+++ src/app/presentation/components/home-components/login/login.component.ts
@@ -29,58 +29,8 @@ export class LoginComponent implements OnInit {
   ngOnInit(): void {
   }
 
-  public externalLoginGoogle = () => {
-    this.service.signInWithGoogle()
-    .then(res => {
-      const user: SocialUser = { ...res };
-      console.log(user);
-      const externalAuth: UserExternalAuth = {
-        provider: user.provider,
-        idToken: user.idToken
-      }
-      this.validateExternalGoogleAuth(externalAuth);
-    }, error => {
-        this.alertService.errorMessage(error)
-    });
-  }
-
-  public externalLoginFB = () => {
-    this.service.signInWithFB()
-    .then(res => {
-      const user: SocialUser = { ...res };
-      console.log(user);
-      const externalAuth: UserExternalAuth = {
-        provider: user.provider,
-        idToken: user.id
-      }
-      this.validateExternalFBAuth(externalAuth);
-    }, error => {
-        this.alertService.errorMessage(error)
-    });
-  }
-
-  private validateExternalGoogleAuth(externalAuth: UserExternalAuth) {
-    this.service.externalGoogleLogin(externalAuth)
-      .subscribe(() => {
-        this.service.isAuthenticatedWithRefreshToken();
-        this.router.navigate(['user/workspaces']);
-      },
-      error => {
-        this.alertService.errorMessage(error);
-        this.service.signOutExternal();
-      });
-  }
-
-  private validateExternalFBAuth(externalAuth: UserExternalAuth) {
-    this.service.externalFBLogin(externalAuth)
-      .subscribe(() => {
-        this.service.isAuthenticatedWithRefreshToken();
-        this.router.navigate(['user/workspaces']);
-      },
-      error => {
-        this.alertService.errorMessage(error);
-        this.service.signOutExternal();
-      });
+  public externalLog(){
+    this.service.externalLoginGoogle();
   }
 
   submit()
