+++ src/app/presentation/components/home-components/registration/registration.component.css
@@ -23,3 +23,45 @@ form.register-form .mat-form-field {
 .back-link a:hover {
     color: #363446;
 }
+.google-btn {
+    margin-top: 20px;
+    margin-left: 90px;
+    width: 300px;
+    height: 38px;
+    background-color: #4285f4;
+    border-radius: 6px;
+    box-shadow: 0 3px 4px 0 rgba(0, 0, 0, 0.25);
+    cursor: pointer;
+}
+.google-btn .google-icon-wrapper {
+    position: absolute;
+    margin-top: 1px;
+    margin-left: 1px;
+    width: 36px;
+    height: 36px;
+    border-radius: 6px;
+    background-color: #fff;
+}
+.google-btn .google-icon {
+    position: absolute;
+    margin-top: 9px;
+    margin-left: 9px;
+    width: 18px;
+    height: 18px;
+}
+.google-btn .btn-text {
+    float: right;
+    margin-top: 9px;
+    color: #fff;
+    font-size: 16px;
+    letter-spacing: 0.6px;
+    font-family: "Roboto";
+    width: 265px;
+    text-align: center;
+}
+.google-btn:hover {
+    box-shadow: 0 0 6px #4285f4;
+}
+.google-btn:active {
+    background: #1669F2;
+}
