+++ src/app/presentation/components/home-components/registration/registration.component.html
@@ -52,6 +52,12 @@
                             </div>
                         </div>
                     </form>
+                    <div class="google-btn">
+                        <div class="google-icon-wrapper">
+                            <img class="google-icon" src="https://upload.wikimedia.org/wikipedia/commons/5/53/Google_%22G%22_Logo.svg"/>
+                        </div>
+                        <p class="btn-text" (click)="externalLog()">Sign in with google</p>
+                    </div>
                 </div>
             </div>
         </div>
